+++ index.html
@@ -228,6 +228,5 @@
 	</template>
 
 	<script type="module" src="js/main.ts"></script>
-	<script type="module" src="js/functions.ts"></script>
 </body>
 </html>
